+++ NBitcoin/DataEncoders/Encoders.cs
@@ -100,13 +100,5 @@ namespace NBitcoin.DataEncoders
 			return new Bech32Encoder(hrp);
 		}
 
-		public static Blech32Encoder Blech32(string hrp)
-		{
-			return new Blech32Encoder(hrp);
-		}
-		public static Blech32Encoder Blech32(byte[] hrp)
-		{
-			return new Blech32Encoder(hrp);
-		}
 	}
 }
